Moved tests to groovy

+++ mobile/src/main/java/com/pablisco/android/sensor/OnShakeListener.java
@@ -2,6 +2,6 @@ package com.pablisco.android.sensor;
 
 public interface OnShakeListener {
 
-    void onShake(double speed);
+    void onShake(float speed);
 
 }
